+++ ElvUI_MerathilisUI/Core/Changelog/6.37.lua
@@ -1,8 +1,17 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[637] = {
-	RELEASE_DATE = "TBD",
-	FIXES = {},
-	NEW = {},
-	IMPROVEMENTS = {},
+	RELEASE_DATE = "31.08.2024",
+	FIXES = {
+		"[MicroBar]: Fix a MicroBar Notification error",
+		"[WorldMap]: Try to avoid a mapCancvas error",
+	},
+	NEW = {
+		"[Event Tracker]: Added Theater Troupe Event",
+	},
+	IMPROVEMENTS = {
+		"[AutoButtons]: Added a lot of new openable items",
+		"[Context Memu]: Some code adjustments and improvements",
+		"[Skins]: More updates to Delves and PlayerSpells skin",
+	},
 }