+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### Changes ###
 
-  * [Armory]: Updates to Armory
-  * [Skins]: Various skin tweaks & updates
-  * [Chat Text]: Sync with ElvUI changes
+  * [MicroBar]: Fix a MicroBar Notification error"
+  * [WorldMap]: Try to avoid a mapCancvas error
+  * [Event Tracker]: Added Theater Troupe Event
+  * [AutoButtons]: Added a lot of new openable items
+  * [Context Memu]: Some code adjustments and improvements
+  * [Skins]: More updates to Delves and PlayerSpells skin